% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wmapshp.R
\name{check_wmapshp}
\alias{check_wmapshp}
\title{Check your SpatialPolygonDataFrame's compatibility with wmap() function.}
\usage{
check_wmapshp(shp)
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame}
}
\value{
TRUE, or function will STOP with print statements about what requirements were not met.
}
\description{
Checks the spatial object you intend to use with the wmap() function
for valid geometry, projection, and the data being a data.table
}
\examples{
check_wmapshp(a_SpatialPolygonsDataFrame)
}

