% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_pallette.R
\name{intensity_pallette}
\alias{intensity_pallette}
\title{Histogram of data using color scheme}
\usage{
intensity_pallette(start, r, gamma, hue_list = c(0.5, 1, 3))
}
\arguments{
\item{start}{numeric start-point for cubehelix color ramp}

\item{r}{number of rotations through cubehelix color ramp}

\item{gamma}{numeric; how light or dark you want it to be 1: normal, higher:darker, lower: lighter}

\item{hue}{numeric;how bright you want the colors to be 1: normal, higher: brighter, lower: more demure}

\item{color_value_breaks}{values to break colors at}

\item{dist_stats}{which distribution to use}
}
\value{
an adapted cubehelix color pallette, increasing in intensity.
}
\description{
This histogram function creates and returns a ggplot object with
a color scale dependent on the X value (default as all colors being grey)
as well as the possibility for vertical lines that describe different
properties of the distribution (mean, median, etc) added on top of the graph.
}
\examples{
intensity_pallette(start=0.5,r=.5,gamma=1)
}

