% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colors.R
\name{plot_colors}
\alias{plot_colors}
\title{Plot colors of a color pallette}
\usage{
plot_colors(color_list, color_list_name = NULL)
}
\arguments{
\item{color_list}{list or vector of valid R colors}

\item{color_list_name}{name of color scheme, optional, default is NULL.}
}
\value{
ggplot of color swatches and optional name of color pallette
}
\description{
This function creates colored swatches of a list
of colors for easy visualization.
}
\examples{
plot_colors(c("red","yellow","blue"),"primary colors")
}
