% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_wpal.R
\name{view_wpal}
\alias{view_wpal}
\title{View a Woodson Pallette color scheme.}
\usage{
view_wpal(color = NULL)
}
\arguments{
\item{color}{the name of the color scheme}
}
\value{
plots of one or all Woodson pallette schemes
}
\description{
This function shows a woodson pallette color scheme when the name is entered. 
 When no name of a wpal() color scheme is entered, all woodson pallette shemes are shown.
}
\examples{
view_wpal()
}

